Добавил экран отображения описания книг

+++ app/src/main/java/com/example/healthapp/model/books/ListPrice.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ListPrice (
-  @SerializedName("amountInMicros" ) var amountInMicros : Int?    = null,
+  @SerializedName("amountInMicros" ) var amountInMicros : Long?    = null,
   @SerializedName("amount" ) var amount : Float?    = null,
   @SerializedName("currencyCode"   ) var currencyCode   : String? = null
